(re) Updating modules, and Node.js
CoffeeScript presented a slight problem, as between 1.4 and 1.6, it stopped inserting a newline
after script-literals ... which breaks the line *following* a script-literal that has a full-line
comment in it. Given that we use that for browserify/tagify, that's icky.

Got a temporary fix in-place, and I've opened [an issue][cs-2996] on CoffeeScript itself.

Also, I'm avoiding updating Browserify for the moment; it seems like the update from Browserify 1 to
Browserify 2 is going to be a snaggy issue, because tagify is no longer supported/maintained.

[cs-2996]: <https://github.com/jashkenas/coffee-script/issues/2996> "Full-line comments in script literals (regression)"