(- re noop) Stop carrying around dead utilities code
This is some great code, and shit I really want to debug, fix up, and publish ... at a later date. I
decided to *not* go with the Function-subclassing API for Paws.js, making all of this code, and the
respective failing tests, unnecessary.

To future-me: remember this is here, in the git history. Come take it, and make a beautiful,
beautiful module. ('=