(!! re API) Rename and refactor terminology
Put a lot of effort into conforming well to the "new terminology". Confusingly, the meaning of "native" is now swapped with its previous meaning, as are "responsible" and "owned". Oh well. Biting the bullet, re-phrasing to be intuitive, and then sticking with my choices.

The only interesting change: the call-pattern for ownership (previously responsibility) has changed, slightly. Expect to use `owned` and `disowned` to return Relations, but `own: yes/no` or `owns = yes/no` to modify the actual values.