(!! meta re) Tool for coverage dynamically
So happy to finally get to use this feature of `coffee-coverage` from 0.4.x. This lets me get rid of my ridiculously fragile `require()` wrapper. PRAISE THE FRUIT PIE or whatever food it is that people use to make fun of theists?

I'm leaving in the `cov_require.js` file for a short while, because I also need to update the matching Primitives files; but that file should disappear entirely, soon.

There's a couple other relevant changes in here (better script-debugging, for one.)